Change consensus condition to full direct propagation of votes
Direct here means that validators do not relay votes from others - they merely
broadcast their original vote on until everyone has seen it. This is done
intentionally to observe basic scaling properties.

To reach the considered consensus condition, we every node needs to broadcast
its original vote to every other node, so (n-1)(n) messages in total.

In the current generation specification, every node currently selected for
broadcast does so for a random selection of its peers. So if we consider a
single message as one of these broadcasts instead, we should need fewer
than (n-1)(n) broadcasts.

However, since the choice of nodes broadcast to is not informed by prior
broadcasts, the effect described in the paragraph above is mitigated. In fact,
from current testing (with up to 40 nodes), it seems that we still require
around (n-1)(n) messages to reach consensus.